+++ packages/kitchensink/jest.globals.js
@@ -1,0 +1,5 @@
+const fastifyTestUtils = require('@forrestjs/service-fastify/test/globals');
+
+module.exports = () => ({
+  ...fastifyTestUtils,
+});