+++ packages/kitchensink/jest.setup.js
@@ -1,0 +1,2 @@
+const global = require('./jest.globals')();
+module.exports = global.awaitTestHealthz;